refactor(script-bindings): modularize + long-tail surface; compact the plan doc
* script_bindings.rs (1741 lines) → script_bindings/{mod,engine,wire,
  tests}.rs (332/650/504/282): mod = public API, cache, contexts,
  marshaling helpers; engine = the full register_fn surface; wire =
  builders/trampolines/option mappers; tests split out. No behavior
  change (contrib 10/10, e2e 1/1, clippy 0).
* Long-tail surface: LookupTokens/LookupCatcode/AssignCatcode/
  LookupMeaning, RefStepID/RefCurrentID, RegisterNamespace/
  RegisterDocumentNamespace/RelaxNGSchema (class-binding essentials),
  DefColumnType (\NC@rewrite@<char> lowering, Rhai expansion body).
* docs: archived the historical M0/M1/M2–M4 logs + the two dated
  critical re-evaluations to docs/archive/SCRIPT_BINDINGS_LOG_2026-06.md
  (653 → 410 live lines); archive README indexed.
